Add a HelloWorldDotNet DLL that we can use for tests to ext. Just a one off, source doesn't matter. Fixed up tool tests to build on the right platform and use the right paths.
