+++ IKVM.sln
@@ -81,6 +81,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IKVM.Tool.Exporter", "src\I
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IKVM.Tool.Exporter.Tests", "src\IKVM.Tool.Exporter.Tests\IKVM.Tool.Exporter.Tests.csproj", "{EFF93C65-3F45-48B6-9178-1E68ADD41974}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IKVM.Tool.Tests", "src\IKVM.Tool.Tests\IKVM.Tool.Tests.csproj", "{E4E6EAB9-2C68-49AC-A212-D464F1BB483B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -203,6 +205,10 @@ Global
 		{EFF93C65-3F45-48B6-9178-1E68ADD41974}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EFF93C65-3F45-48B6-9178-1E68ADD41974}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EFF93C65-3F45-48B6-9178-1E68ADD41974}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E4E6EAB9-2C68-49AC-A212-D464F1BB483B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E4E6EAB9-2C68-49AC-A212-D464F1BB483B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E4E6EAB9-2C68-49AC-A212-D464F1BB483B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E4E6EAB9-2C68-49AC-A212-D464F1BB483B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
