+++ ext/netcoreapp3.1/HelloWorldDotNet.deps.json
@@ -1,0 +1,23 @@
+{
+  "runtimeTarget": {
+    "name": ".NETCoreApp,Version=v3.1",
+    "signature": ""
+  },
+  "compilationOptions": {},
+  "targets": {
+    ".NETCoreApp,Version=v3.1": {
+      "HelloWorldDotNet/1.0.0": {
+        "runtime": {
+          "HelloWorldDotNet.dll": {}
+        }
+      }
+    }
+  },
+  "libraries": {
+    "HelloWorldDotNet/1.0.0": {
+      "type": "project",
+      "serviceable": false,
+      "sha512": ""
+    }
+  }
+}
