+++ src/IKVM.Tool.Tests/IKVM.Tool.Tests.csproj
@@ -1,0 +1,101 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFrameworks>net461;netcoreapp3.1;net5.0;net6.0</TargetFrameworks>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="FluentAssertions" Version="6.7.0" />
+        <PackageReference Include="Microsoft.CodeCoverage" Version="17.2.0" />
+        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.2.0" />
+        <PackageReference Include="MSTest.TestAdapter" Version="2.2.10" />
+        <PackageReference Include="MSTest.TestFramework" Version="2.2.10" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\IKVM.Tool\IKVM.Tool.csproj" />
+    </ItemGroup>
+
+    <ItemGroup>
+        <None Include="..\..\ext\**\*" LinkBase="ext">
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </None>
+    </ItemGroup>
+
+    <ItemGroup Condition=" '$(TargetFramework)' != '' ">
+        <IncludeProjectReference Include="..\IKVM.Runtime\IKVM.Runtime.csproj">
+            <SetTargetFramework>TargetFramework=net461</SetTargetFramework>
+            <IncludeTargetPath>lib\net461\</IncludeTargetPath>
+        </IncludeProjectReference>
+        <IncludeProjectReference Include="..\IKVM.Runtime.JNI\IKVM.Runtime.JNI.csproj">
+            <SetTargetFramework>TargetFramework=net461</SetTargetFramework>
+            <IncludeTargetPath>lib\net461\</IncludeTargetPath>
+        </IncludeProjectReference>
+        <IncludeProjectReference Include="..\IKVM.Java\IKVM.Java.msbuildproj">
+            <SetTargetFramework>TargetFramework=net461</SetTargetFramework>
+            <IncludeTargetPath>lib\net461\</IncludeTargetPath>
+        </IncludeProjectReference>
+        <IncludeProjectReference Include="..\IKVM.Runtime\IKVM.Runtime.csproj">
+            <SetTargetFramework>TargetFramework=netcoreapp3.1</SetTargetFramework>
+            <IncludeTargetPath>lib\netcoreapp3.1\</IncludeTargetPath>
+        </IncludeProjectReference>
+        <IncludeProjectReference Include="..\IKVM.Runtime.JNI\IKVM.Runtime.JNI.csproj">
+            <SetTargetFramework>TargetFramework=netcoreapp3.1</SetTargetFramework>
+            <IncludeTargetPath>lib\netcoreapp3.1\</IncludeTargetPath>
+        </IncludeProjectReference>
+        <IncludeProjectReference Include="..\IKVM.Java\IKVM.Java.msbuildproj">
+            <SetTargetFramework>TargetFramework=netcoreapp3.1</SetTargetFramework>
+            <IncludeTargetPath>lib\netcoreapp3.1\</IncludeTargetPath>
+        </IncludeProjectReference>
+    </ItemGroup>
+
+    <ItemGroup Condition=" '$(TargetFramework)' != '' ">
+
+        <PublishProjectReference Include="..\ikvm\ikvm.csproj">
+            <SetTargetFramework>TargetFramework=net461</SetTargetFramework>
+            <PublishTargetPath>ikvm\net461\any</PublishTargetPath>
+        </PublishProjectReference>
+        <PublishProjectReference Include="..\ikvmc\ikvmc.csproj">
+            <SetTargetFramework>TargetFramework=net461</SetTargetFramework>
+            <PublishTargetPath>ikvmc\net461\any</PublishTargetPath>
+        </PublishProjectReference>
+        <PublishProjectReference Include="..\ikvmstub\ikvmstub.csproj">
+            <SetTargetFramework>TargetFramework=net461</SetTargetFramework>
+            <PublishTargetPath>ikvmstub\net461\any</PublishTargetPath>
+        </PublishProjectReference>
+
+        <PublishProjectReference Include="..\ikvm\ikvm.csproj">
+            <SetTargetFramework>TargetFramework=netcoreapp3.1</SetTargetFramework>
+            <PublishTargetPath>ikvm\netcoreapp3.1\win7-x64</PublishTargetPath>
+            <PublishProperties>RuntimeIdentifier=win7-x64</PublishProperties>
+        </PublishProjectReference>
+        <PublishProjectReference Include="..\ikvmc\ikvmc.csproj">
+            <SetTargetFramework>TargetFramework=netcoreapp3.1</SetTargetFramework>
+            <PublishTargetPath>ikvmc\netcoreapp3.1\win7-x64</PublishTargetPath>
+            <PublishProperties>RuntimeIdentifier=win7-x64</PublishProperties>
+        </PublishProjectReference>
+        <PublishProjectReference Include="..\ikvmstub\ikvmstub.csproj">
+            <SetTargetFramework>TargetFramework=netcoreapp3.1</SetTargetFramework>
+            <PublishTargetPath>ikvmstub\netcoreapp3.1\win7-x64</PublishTargetPath>
+            <PublishProperties>RuntimeIdentifier=win7-x64</PublishProperties>
+        </PublishProjectReference>
+
+        <PublishProjectReference Include="..\ikvm\ikvm.csproj">
+            <SetTargetFramework>TargetFramework=netcoreapp3.1</SetTargetFramework>
+            <PublishTargetPath>ikvm\netcoreapp3.1\linux-x64</PublishTargetPath>
+            <PublishProperties>RuntimeIdentifier=linux-x64</PublishProperties>
+        </PublishProjectReference>
+        <PublishProjectReference Include="..\ikvmc\ikvmc.csproj">
+            <SetTargetFramework>TargetFramework=netcoreapp3.1</SetTargetFramework>
+            <PublishTargetPath>ikvmc\netcoreapp3.1\linux-x64</PublishTargetPath>
+            <PublishProperties>RuntimeIdentifier=linux-x64</PublishProperties>
+        </PublishProjectReference>
+        <PublishProjectReference Include="..\ikvmstub\ikvmstub.csproj">
+            <SetTargetFramework>TargetFramework=netcoreapp3.1</SetTargetFramework>
+            <PublishTargetPath>ikvmstub\netcoreapp3.1\linux-x64</PublishTargetPath>
+            <PublishProperties>RuntimeIdentifier=linux-x64</PublishProperties>
+        </PublishProjectReference>
+
+    </ItemGroup>
+
+</Project>
